Fix all instances of e.printStackTrace()
Not terribly important, but low hanging fruit now that Timber is in
place.

Note for reviewers: I'm not sure we want all of those in Crashlytics,
please consider that. But maybe too much data in Crashlytics is better
than too little? In that case these should probably be dealt with on a
case-by-case basis.